requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
asyncpg
psycopg2
python-dotenv
Cython
graphql
strawberry-graphql[fastapi]
python-jose[cryptography]
passlib[bcrypt]